calculator double

+++ Calculator/Program.cs
@@ -7,12 +7,12 @@
             string input1 = Console.ReadLine();
         Console.WriteLine("enter second number:");
             string input2 = Console.ReadLine();
-        int number1 = int.Parse(input1);
-        int number2 = int.Parse(input2);
-        int sum1 = number1 + number2;
-        int sub1 = number1 - number2;
-        int mul1 = number1 * number2;
-        int div1 = number1 / number2;
+        double number1 = Double.Parse(input1);
+        double number2 = Double.Parse(input2);
+        double sum1 = number1 + number2;
+        double sub1 = number1 - number2;
+        double mul1 = number1 * number2;
+        double div1 = number1 / number2;
         Console.WriteLine("the sum of these numbers is:" + ' ' + sum1);
         Console.WriteLine("the substraction of these numbers is:" + ' ' + sub1);
         Console.WriteLine("the multiplication of these numbers is:" + ' ' + mul1);
